+++ src/iGL.Engine/GameComponents/LightComponent.cs
@@ -19,7 +19,7 @@ namespace iGL.Engine
            
         }
 
-        public override void Tick(double timeElapsed)
+        public override void Tick(float timeElapsed)
         {
             
         }
